requirements.txt
aiohttp==3.8.5
Flask==2.2.2
python-dotenv==1.0.0
slack-sdk==3.21.3
slackeventsapi==3.0.1aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.2.0
click==8.1.6
dnspython==2.4.1
Flask==2.2.2
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
pyee==11.0.0
pymongo==4.4.1
python-dotenv==1.0.0
slack-sdk==3.21.3
slackeventsapi==3.0.1
typing-extensions==4.7.1
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2
